+++ spec/spec_helper.rb
@@ -5,6 +5,8 @@ ENV["RAILS_ENV"] ||= "test"
 require File.expand_path("dummy/config/environment", __dir__)
 require "rspec/rails"
 
+Dir[File.expand_path("support/**/*.rb", __dir__)].each { |f| require f }
+
 RSpec.configure do |config|
   config.expect_with :rspec do |expectations|
     expectations.include_chain_clauses_in_custom_matcher_descriptions = true
@@ -20,8 +22,10 @@ RSpec.configure do |config|
   Kernel.srand config.seed
 
   # Reset StandardHealth config between examples so registered checks /
-  # env_specs don't leak.
+  # env_specs / parent controllers don't leak.
   config.before(:each) do
     StandardHealth.reset_config!
+    StandardHealth.reset_parent_controller!
+    StandardHealth.reset_diagnostics_parent_controller!
   end
 end